pyproject.toml
[project]
name = "threeb1b_hand"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.14.0",
    "cryptography>=41.0.0",
    "pillow>=12.0.0",
    "playwright>=1.57.0",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]
